move merkle tree membership code to barretenberg_wasm
